EXP-2012: Move tests for enrollment into their own test files
Make previous enrollment time const public

Remove unused import

Move unit tests to their own folder under source. Remove all the unnecessary pub's now that the scopes are happy

Limit public functions to the crate only
